# Compactor pooling changes